Flip the ScrollVar logic to a positive offset.
That way if you directly set a value, it will be the one reported back
to you.

+++ source/OutfitterPanel.cpp
@@ -241,7 +241,7 @@ int OutfitterPanel::DetailWidth() const
 
 
 
-int OutfitterPanel::DrawDetails(const Point &center)
+double OutfitterPanel::DrawDetails(const Point &center)
 {
 	string selectedItem = "Nothing Selected";
 	const Font &font = FontSet::Get(14);
